+++ src/main/java/com/examen_ad/academia/repository/ProfesorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Integer> {
 }
